+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 ChangeLog
 =========
 
+0.0.1.4 (2016-08-24)
+--------------------
+
+- Added moving progress bar
+- Change moving to copy-delete. see .. _Issue #1: https://github.com/aviadlevy/ktuvitDownloader/issues/1
+- Now logger is rotating (on 1MB)
+
 0.0.1.2-3 (2016-08-23)
 ----------------------